Two major capabilities have been merged into the main branch from liquidTransportDevelop branch
The Exchange_current_reaction_ratecoeff_type method for
formulating reaction rate constants for exchange current
reactions has been ported into the main branch.

Fixed the InterfaceKinetics object so that we can
specify that a phase doesn't exist. If a phase doesn't exist,
species in that phase can't be reactants in the phase. Before,
this fix, they could be !?!!
This was because InterfaceKinetics objects use ThermoPhase
objects, which are intrinsic only. We need an extrinsic notion 
that is in addition to ThermoPhase info in order to get this right.




git-svn-id: https://cantera.googlecode.com/svn/cantera18/trunk@462 02a645c2-efd0-11dd-984d-ab748d24aa7e