Fixed the order of include and library directories when compiling
Include directories in the Cantera source tree should be listed first so that
they take precedence over headers from other installed copies of Cantera that
might end up on the include search path. This was potentially a problem when
Cantera was installed in the same location (e.g. /usr/local) as one of its
dependencies (e.g. Sundials).

The same logic applies to directories on the library link path.

git-svn-id: https://cantera.googlecode.com/svn/cantera/branches/2.0@1777 02a645c2-efd0-11dd-984d-ab748d24aa7e